pyproject.toml
[tool.poetry]
name = "mediacollector"
version = "0.1.0"
description = ""
authors = ["Ikey168"]
license = "MIT License"
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
numpy = "^1.26.2"
config = "^0.5.1"
requests = "^2.31.0"
tqdm = "^4.66.1"
newspaper3k = "^0.2.8"
pytube = "^15.0.0"
openai-whisper = "^20231117"
praw = "^7.7.1"
feedparser = "^6.0.11"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
